--- a/Cargo.toml
+++ b/Cargo.toml
@@ -66,7 +66,6 @@ nu-test-support = {version = "0.25.2", path = "./crates/nu-test-support"}
 ctrlc-support = ["nu-cli/ctrlc", "nu-command/ctrlc"]
 directories-support = ["nu-cli/directories", "nu-cli/dirs", "nu-command/directories", "nu-command/dirs", "nu-data/directories", "nu-data/dirs", "nu-engine/dirs"]
 ptree-support = ["nu-cli/ptree", "nu-command/ptree"]
-rich-benchmark = ["nu-cli/rich-benchmark", "nu-command/rich-benchmark"]
 rustyline-support = ["nu-cli/rustyline-support", "nu-command/rustyline-support"]
 term-support = ["nu-cli/term", "nu-command/term"]
 uuid-support = ["nu-cli/uuid_crate", "nu-command/uuid_crate"]
@@ -87,7 +86,6 @@ default = [
   "match",
   "post",
   "fetch",
-  "rich-benchmark",
   "zip-support",
 ]
 extra = ["default", "binaryview", "tree", "clipboard-cli", "trash-support", "start", "bson", "sqlite", "s3", "chart", "xpath", "selector"]