--- a/crates/nu-cli/Cargo.toml
+++ b/crates/nu-cli/Cargo.toml
@@ -55,7 +55,6 @@ futures-util = "0.3.8"
 futures_codec = "0.4.1"
 getset = "0.1.1"
 glob = "0.3.0"
-heim = { version = "0.1.0-rc.1", optional = true }
 htmlescape = "0.3.1"
 ical = "0.7.0"
 ichwh = { version = "0.3.4", optional = true }
@@ -98,7 +97,6 @@ titlecase = "1.0"
 toml = "0.5.6"
 trash = { version = "1.2.0", optional = true }
 unicode-segmentation = "1.6.0"
-uom = { version = "0.30.0", features = ["f64", "try-from"] }
 url = "2.1.1"
 uuid_crate = { package = "uuid", version = "0.8.1", features = ["v4"], optional = true }
 which = { version = "4.0.2", optional = true }
@@ -130,7 +128,6 @@ quickcheck_macros = "0.9.1"
 [features]
 default = ["shadow-rs"]
 clipboard-cli = ["arboard"]
-rich-benchmark = ["heim"]
 rustyline-support = ["rustyline", "nu-engine/rustyline-support"]
 stable = []
 trash-support = ["trash"]