--- a/crates/nu-command/src/commands/version.rs
+++ b/crates/nu-command/src/commands/version.rs
@@ -170,10 +170,10 @@ fn features_enabled() -> Vec<String> {
         names.push("ptree".to_string());
     }
 
-    #[cfg(feature = "rich-benchmark")]
-    {
-        names.push("rich-benchmark".to_string());
-    }
+    // #[cfg(feature = "rich-benchmark")]
+    // {
+    //     names.push("rich-benchmark".to_string());
+    // }
 
     #[cfg(feature = "rustyline-support")]
     {