--- a/crates/nu_plugin_ps/Cargo.toml
+++ b/crates/nu_plugin_ps/Cargo.toml
@@ -19,10 +19,6 @@ num-bigint = "0.3.0"
 
 futures = {version = "0.3.5", features = ["compat", "io-compat"]}
 futures-timer = "3.0.2"
-
-[dependencies.heim]
-default-features = false
-features = ["process"]
-version = "0.1.0-rc.1"
+sysinfo = "0.15.9"
 
 [build-dependencies]