This moves us onto using sysinfo instead of heim+uom. There are a few changes as a result:

* There aren't 1:1 equivalents for heim in sysinfo, which means some of the features we used to have aren't supported. Notably now missing: battery info, some of the system info fields, and maybe 1(?) of the ps fields.

* The reported processor memory usage differs. I noticed that current `nu` also differs from other process monitor tools I use, so perhaps different apps just count sizes differently.

* We also lose the fancy `benchmark` output because it relied on uom.

Performance between the two seems relatively similar.

Compile times improve noticeably. Here's a before/after `cargo bloat` run (in total, a `cargo build --all --features=extra` build drops from 5 mins 30 sec to 5 mins for me):

## Before
```
  Time Crate
61.78s nu_command
40.93s uom
 9.85s syn
 8.95s nu_engine
 8.10s nu
 7.59s clap
```
## After 
```
  Time Crate
72.37s nu_command
10.75s nu_engine
10.61s syn
 8.65s nu_protocol
 8.54s clap
 8.42s nu
 8.31s nu_plugin
```