@tshepang 2021-01-27
Are compile times the primary driver for this change? I ask because no other motivation is given, considering the regression in features.